Rename styles
