+++ src/components/contact/ContactDetails.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from '@reach/router';
 
-import '../ComponentStyle.css';
+import '../style.css';
 
 const ContactDetails = (props) => {     
     return (
